+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## [0.1.31] - 2026-08-01
+
+## Other
+
+- affinamenti pre regex - (Release 0.1.31)
+- affinamenti pre regex - (Release 0.1.31)
+
 ## [0.1.31] - 2026-07-31
 
 ## Other